+++ src/tkc/slist.h
@@ -23,6 +23,7 @@
 #define TK_SLIST_H
 
 #include "tkc/types_def.h"
+#include "tkc/mem_allocator.h"
 
 BEGIN_C_DECLS
 
@@ -90,6 +91,12 @@ typedef struct _slist_t {
    * 元素比较函数。
    */
   tk_compare_t compare;
+
+  /**
+   * @property {mem_allocator_t*} node_allocator
+   * 节点内存分配器。
+   */
+  mem_allocator_t* node_allocator;
 } slist_t;
 
 /**
@@ -287,6 +294,16 @@ int32_t slist_count(slist_t* slist, void* ctx);
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
 ret_t slist_reverse(slist_t* slist);
+
+/**
+ * @method slist_set_node_allocator
+ * 设置节点内存分配器。
+ * @param {slist_t*} slist 单向链表对象。
+ * @param {mem_allocator_t*} allocator 内存分配器对象。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t slist_set_node_allocator(slist_t* slist, mem_allocator_t* allocator);
 
 /**
  * @method slist_deinit